chore: format using forge fmt

+++ script/ValidateSPV.s.sol
@@ -1,10 +1,12 @@
 pragma solidity ^0.8.4;
 
-/** @title ValidateSPV*/
-/** @author Summa (https://summa.one) */
-
+/**
+ * @title ValidateSPV
+ */
+/**
+ * @author Summa (https://summa.one)
+ */
 import {ValidateSPV} from "../src/ValidateSPV.sol";
-
 
 contract ValidateSPVScript {
     function getErrBadLength() public pure returns (uint256) {
@@ -25,12 +27,7 @@ contract ValidateSPVScript {
     /// @param _proof           The proof (concatenated LE hashes)
     /// @param _index           The proof index
     /// @return                 true if fully valid, false otherwise
-    function prove(
-        bytes32 _txid,
-        bytes32 _merkleRoot,
-        bytes memory _proof,
-        uint _index
-    ) public returns (bool) {
+    function prove(bytes32 _txid, bytes32 _merkleRoot, bytes memory _proof, uint256 _index) public returns (bool) {
         return ValidateSPV.prove(_txid, _merkleRoot, _proof, _index);
     }
 
@@ -41,12 +38,10 @@ contract ValidateSPVScript {
     /// @param _vout        Raw bytes length-prefixed output vector
     /// @ param _locktime   4-byte tx locktime
     /// @return             32-byte transaction id, little endian
-    function calculateTxId(
-        bytes4 _version,
-        bytes memory _vin,
-        bytes memory _vout,
-        bytes4 _locktime
-    ) public returns (bytes32) {
+    function calculateTxId(bytes4 _version, bytes memory _vin, bytes memory _vout, bytes4 _locktime)
+        public
+        returns (bytes32)
+    {
         return ValidateSPV.calculateTxId(_version, _vin, _vout, _locktime);
     }
